Distance Measurement Example
this c++ file is a teset code that measures distancee from the ultrasonic sensor and prints it in mm and cm. However, when running this using the wio terminal and the ultrasonic sensor, it did not print the distance of the objects accurately and the output was zero all the time. I am commiting this so that everybody can work on this and help with the problem encountered