#55 Start button Loading scene
I used the loading scen that Kaisa had made, made some changes to the way the lastCommandReceived when the user toggels over the button. I wanted to render the loading scene on the button, but I could not find a way for that as it would cause the test to fail. This is the best I came up with